requirements.txt
django-taggit
django-xadmin
